refactor: replace useSuspenseRole with PageMonoview
