+++ src/app/(dashboard)/[guildUrlName]/hooks/usePageMonoview.ts
@@ -1,5 +1,5 @@
 import { pageMonoviewOptions } from "@/lib/options";
-import { useQuery } from "@tanstack/react-query";
+import { useQuery, useSuspenseQuery } from "@tanstack/react-query";
 import { useParams } from "next/navigation";
 
 export const usePageMonoview = () => {
@@ -8,7 +8,23 @@ export const usePageMonoview = () => {
     guildUrlName: string;
   }>();
 
-  return useQuery(
+  // biome-ignore lint/suspicious/noExplicitAny: <explanation>
+  return useQuery<any>(
+    pageMonoviewOptions({
+      guildIdLike: guildUrlName,
+      pageIdLike: pageUrlNameParam,
+    }),
+  );
+};
+
+export const usePageMonoviewSuspense = () => {
+  const { pageUrlName: pageUrlNameParam, guildUrlName } = useParams<{
+    pageUrlName: string;
+    guildUrlName: string;
+  }>();
+
+  // biome-ignore lint/suspicious/noExplicitAny: <explanation>
+  return useSuspenseQuery<any>(
     pageMonoviewOptions({
       guildIdLike: guildUrlName,
       pageIdLike: pageUrlNameParam,
